+++ stroom-security-server/src/main/java/stroom/security/server/AuthenticationServiceClients.java
@@ -108,7 +108,7 @@ public class AuthenticationServiceClients {
         }
     }
 
-    private String createTokenForUser(String userId) throws ApiException {
+    String createTokenForUser(String userId) throws ApiException {
         CreateTokenRequest createTokenRequest = new CreateTokenRequest();
         createTokenRequest.setEnabled(true);
         createTokenRequest.setTokenType("api");